refactor(android): split session detail responsibilities
